// frontend/src/api/generated/models/import-node-type.ts
/* tslint:disable */
/* eslint-disable */
/**
 * linkweave-api API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0-SNAPSHOT
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ImportNodeType = {
    Folder: 'FOLDER',
    Bookmark: 'BOOKMARK'
} as const;
export type ImportNodeType = typeof ImportNodeType[keyof typeof ImportNodeType];


export function instanceOfImportNodeType(value: any): boolean {
    for (const key in ImportNodeType) {
        if (Object.prototype.hasOwnProperty.call(ImportNodeType, key)) {
            if (ImportNodeType[key as keyof typeof ImportNodeType] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ImportNodeTypeFromJSON(json: any): ImportNodeType {
    return ImportNodeTypeFromJSONTyped(json, false);
}

export function ImportNodeTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ImportNodeType {
    return json as ImportNodeType;
}

export function ImportNodeTypeToJSON(value?: ImportNodeType | null): any {
    return value as any;
}

export function ImportNodeTypeToJSONTyped(value: any, ignoreDiscriminator: boolean): ImportNodeType {
    return value as ImportNodeType;
}